Fix long long int printf format argument for amd64 platform
git-svn-id: https://devel.aqbanking.de/svn/aqbanking/trunk@982 5c42a225-8b10-0410-9873-89b7810ad06e

+++ src/libs/aqbanking/wcb.c
@@ -82,7 +82,7 @@ GWEN_WAITCALLBACK_RESULT AB_WaitCallback_CheckAbort(GWEN_WAITCALLBACK *ctx,
       wcb->lastProgress=GWEN_WaitCallback_GetProgressPos(ctx);
       DBG_NOTICE(AQBANKING_LOGDOMAIN,
 		 "Progress changed to %lld",
-		 wcb->lastProgress);
+		 (long long int)wcb->lastProgress);
       rv=AB_Banking_ProgressAdvance(wcb->banking, 0, wcb->lastProgress);
     }
   }
